refactor(eval): move prebuilt index staging
