+++ scripts/agent_compile/run_synthesis_sweep.py
@@ -43,17 +43,17 @@ _PROJECT_ROOT = Path(__file__).resolve().parents[2]
 if str(_PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(_PROJECT_ROOT))
 
+from codeminer.eval.agent_runner.prebuilt import (  # noqa: E402
+    has_full_indexes,
+    repo_path_for,
+    stage_prebuilt_indexes,
+)
 from scripts.agent_compile.lib.config import SweepConfig  # noqa: E402
 from scripts.agent_compile.lib.harness import (  # noqa: E402
     build_symbol_span_index,
     load_full_contexts,
     run_cell,
     slug,
-)
-from scripts.agent_compile.lib.prebuilt import (  # noqa: E402
-    has_full_indexes,
-    repo_path_for,
-    stage_prebuilt_indexes,
 )
 
 # language_group / config -> SessionContext primary_language key.
