+++ scripts/retrieval_ablation/graphrag_retrieve.py
@@ -84,9 +84,12 @@ def _node_payload(nodes: List[Any], limit: int = 30) -> List[Dict[str, Any]]:
 def _retrieve(instance_id: str, cfg: Any, prebuilt_dir: str, cache_root: str):
     """Run the GraphRAG composer once; return nodes + target metadata."""
     from codeminer.agent.skills.registry import SkillRegistry
+    from codeminer.eval.agent_runner.prebuilt import (
+        repo_path_for,
+        stage_prebuilt_indexes,
+    )
     from codeminer.eval.retrieval_eval import collect_target_blocks, collect_targets
     from scripts.agent_compile.lib.harness import load_dataset_rows, load_full_contexts
-    from scripts.agent_compile.lib.prebuilt import repo_path_for, stage_prebuilt_indexes
 
     rows_by_id, eval_lookup = load_dataset_rows(cfg)
     row = rows_by_id[instance_id]
