+++ test/scripts/test_graphrag_retrieve.py
@@ -72,7 +72,7 @@ def test_retrieve_loads_contexts_from_prebuilt_repo_not_staged_cache(
         fake_load_dataset_rows,
     )
     monkeypatch.setattr(
-        "scripts.agent_compile.lib.prebuilt.stage_prebuilt_indexes",
+        "codeminer.eval.agent_runner.prebuilt.stage_prebuilt_indexes",
         fake_stage_prebuilt_indexes,
     )
     monkeypatch.setattr(
